requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0
pillow-heif>=0.13.0
tqdm>=4.65.0
